start collecting division by zero tests
