+++ pandas/tests/indexes/test_numeric.py
@@ -162,48 +162,6 @@ class Numeric(Base):
         for r, e in zip(result, expected):
             tm.assert_series_equal(r, e)
 
-    def test_div_zero(self, zero):
-        idx = self.create_index()
-
-        expected = Index([np.nan, np.inf, np.inf, np.inf, np.inf],
-                         dtype=np.float64)
-        result = idx / zero
-        tm.assert_index_equal(result, expected)
-        ser_compat = Series(idx).astype('i8') / np.array(zero).astype('i8')
-        tm.assert_series_equal(ser_compat, Series(result))
-
-    def test_floordiv_zero(self, zero):
-        idx = self.create_index()
-        expected = Index([np.nan, np.inf, np.inf, np.inf, np.inf],
-                         dtype=np.float64)
-
-        result = idx // zero
-        tm.assert_index_equal(result, expected)
-        ser_compat = Series(idx).astype('i8') // np.array(zero).astype('i8')
-        tm.assert_series_equal(ser_compat, Series(result))
-
-    def test_mod_zero(self, zero):
-        idx = self.create_index()
-
-        expected = Index([np.nan, np.nan, np.nan, np.nan, np.nan],
-                         dtype=np.float64)
-        result = idx % zero
-        tm.assert_index_equal(result, expected)
-        ser_compat = Series(idx).astype('i8') % np.array(zero).astype('i8')
-        tm.assert_series_equal(ser_compat, Series(result))
-
-    def test_divmod_zero(self, zero):
-        idx = self.create_index()
-
-        exleft = Index([np.nan, np.inf, np.inf, np.inf, np.inf],
-                       dtype=np.float64)
-        exright = Index([np.nan, np.nan, np.nan, np.nan, np.nan],
-                        dtype=np.float64)
-
-        result = divmod(idx, zero)
-        tm.assert_index_equal(result[0], exleft)
-        tm.assert_index_equal(result[1], exright)
-
     def test_explicit_conversions(self):
 
         # GH 8608
