+++ pandas/tests/indexes/timedeltas/test_arithmetic.py
@@ -307,55 +307,6 @@ class TestTimedeltaIndexArithmetic(object):
         tm.assert_index_equal(rng, expected)
 
     # -------------------------------------------------------------
-    # __add__/__sub__ with ndarray[datetime64] and ndarray[timedelta64]
-
-    def test_tdi_sub_dt64_array(self):
-        dti = pd.date_range('2016-01-01', periods=3)
-        tdi = dti - dti.shift(1)
-        dtarr = dti.values
-
-        with pytest.raises(TypeError):
-            tdi - dtarr
-
-        # TimedeltaIndex.__rsub__
-        expected = pd.DatetimeIndex(dtarr) - tdi
-        result = dtarr - tdi
-        tm.assert_index_equal(result, expected)
-
-    def test_tdi_add_dt64_array(self):
-        dti = pd.date_range('2016-01-01', periods=3)
-        tdi = dti - dti.shift(1)
-        dtarr = dti.values
-
-        expected = pd.DatetimeIndex(dtarr) + tdi
-        result = tdi + dtarr
-        tm.assert_index_equal(result, expected)
-        result = dtarr + tdi
-        tm.assert_index_equal(result, expected)
-
-    def test_tdi_add_td64_array(self):
-        dti = pd.date_range('2016-01-01', periods=3)
-        tdi = dti - dti.shift(1)
-        tdarr = tdi.values
-
-        expected = 2 * tdi
-        result = tdi + tdarr
-        tm.assert_index_equal(result, expected)
-        result = tdarr + tdi
-        tm.assert_index_equal(result, expected)
-
-    def test_tdi_sub_td64_array(self):
-        dti = pd.date_range('2016-01-01', periods=3)
-        tdi = dti - dti.shift(1)
-        tdarr = tdi.values
-
-        expected = 0 * tdi
-        result = tdi - tdarr
-        tm.assert_index_equal(result, expected)
-        result = tdarr - tdi
-        tm.assert_index_equal(result, expected)
-
-    # -------------------------------------------------------------
 
     def test_subtraction_ops(self):
         # with datetimes/timedelta and tdi/dti
